+++ tests/Database/Functional/Driver/Common/Schema/AlterColumnTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Cycle\Database\Tests\Functional\Driver\Common\Schema;
 
 use Cycle\Database\Driver\Handler;
+use Cycle\Database\Exception\SchemaException;
 use Cycle\Database\Schema\AbstractColumn;
 use Cycle\Database\Schema\AbstractTable;
 use Cycle\Database\Tests\Functional\Driver\Common\BaseTest;
@@ -113,7 +114,7 @@ abstract class AlterColumnTest extends BaseTest
 
     public function testColumnSizeException(): void
     {
-        $this->expectException(\Cycle\Database\Exception\SchemaException::class);
+        $this->expectException(SchemaException::class);
         $schema = $this->sampleSchema('table');
         $this->assertTrue($schema->exists());
 
@@ -132,6 +133,26 @@ abstract class AlterColumnTest extends BaseTest
         $this->assertTrue(true);
     }
 
+    public function testDatetimeColumnSizeException(): void
+    {
+        $this->expectException(SchemaException::class);
+        $schema = $this->sampleSchema('table');
+        $this->assertTrue($schema->exists());
+
+        $schema->datetime('datetime', -1);
+        $schema->save();
+    }
+
+    public function testDatetimeColumnSize2Exception(): void
+    {
+        $this->expectException(SchemaException::class);
+        $schema = $this->sampleSchema('table');
+        $this->assertTrue($schema->exists());
+
+        $schema->datetime('datetime', 7);
+        $schema->save();
+    }
+
     public function testChangeSize(): void
     {
         $schema = $this->sampleSchema('table');
@@ -144,6 +165,20 @@ abstract class AlterColumnTest extends BaseTest
 
         $this->assertSameAsInDB($schema);
         $this->assertSame(100, $this->fetchSchema($schema)->column('first_name')->getSize());
+    }
+
+    public function testChangeDatetimeSize(): void
+    {
+        $schema = $this->sampleSchema('table');
+        $this->assertTrue($schema->exists());
+
+        $this->assertSame(0, $this->fetchSchema($schema)->column('datetime')->getSize());
+
+        $schema->datetime->string(6);
+        $schema->save();
+
+        $this->assertSameAsInDB($schema);
+        $this->assertSame(6, $this->fetchSchema($schema)->column('datetime')->getSize());
     }
 
     public function testDecimalSizes(): void
@@ -163,7 +198,7 @@ abstract class AlterColumnTest extends BaseTest
 
     public function testDecimalSizesException(): void
     {
-        $this->expectException(\Cycle\Database\Exception\SchemaException::class);
+        $this->expectException(SchemaException::class);
         $schema = $this->sampleSchema('table');
         $this->assertTrue($schema->exists());